+++ libexec/scoop-cat.ps1
@@ -2,7 +2,8 @@
 # Summary: Show content of specified manifest(s).
 #
 # Options:
-#   -h, --help      Show help for this command.
+#   -h, --help                  Show help for this command.
+#   -f, --format <json|yaml>    Show manifest in specific format. Json will be considered as default when this parameter is not provided.
 
 'core', 'getopt', 'help', 'Helpers', 'install', 'manifest' | ForEach-Object {
     . (Join-Path $PSScriptRoot "..\lib\$_.ps1")
@@ -10,41 +11,37 @@
 
 $ExitCode = 0
 $Problems = 0
-$Options, $Applications, $_err = getopt $args
+$Options, $Applications, $_err = getopt $args 'f:' 'format='
 
 if ($_err) { Stop-ScoopExecution -Message "scoop cat: $_err" -ExitCode 2 }
 if (!$Applications) { Stop-ScoopExecution -Message 'Parameter <APP> missing' -Usage (my_usage) }
 
+$Format = $Options.f, $Options.format, 'json' | Where-Object { ! [String]::IsNullOrEmpty($_) } | Select-Object -First 1
+if ($Format -notin $ALLOWED_MANIFEST_EXTENSION) { Stop-ScoopExecution -Message "Format '$Format' is not supported" -ExitCode 2 }
+
 foreach ($app in $Applications) {
-    # Prevent leaking variables from previous iteration
-    $cleanAppName = $bucket = $version = $appName = $manifest = $foundBucket = $url = $null
-
-    # TODO: Adopt Resolve-ManifestInformation
-    $cleanAppName, $bucket, $version = parse_app $app
-    $appName, $manifest, $foundBucket, $url = Find-Manifest $cleanAppName $bucket
-    if ($null -eq $bucket) { $bucket = $foundBucket }
-
-    # Handle potential use case, which should not appear, but just in case
-    # If parsed name/bucket is not same as the provided one
-    if ((!$url) -and (($cleanAppName -ne $appName) -or ($bucket -ne $foundBucket))) {
-        debug $bucket
-        debug $cleanAppName
-        debug $foundBucket
-        debug $appName
-        Write-UserMessage -Message 'Found application name or bucket is not same as requested' -Err
+    $resolved = $null
+    try {
+        $resolved = Resolve-ManifestInformation -ApplicationQuery $app
+    } catch {
         ++$Problems
+
+        $title, $body = $_.Exception.Message -split '\|-'
+        if (!$body) { $body = $title }
+        Write-UserMessage -Message $body -Err
+        debug $_.InvocationInfo
+        if ($title -ne 'Ignore' -and ($title -ne $body)) { New-IssuePrompt -Application $appName -Bucket $bucket -Title $title -Body $body }
+
         continue
     }
 
-    if ($manifest) {
-        Write-UserMessage -Message "Showing manifest for $app" -Color 'Green'
+    debug $resolved
 
-        # TODO: YAML
-        $manifest | ConvertToPrettyJson | Write-UserMessage -Output
-    } else {
-        Write-UserMessage -Message "Manifest for $app not found" -Err
-        ++$Problems
-        continue
+    $output = $resolved.ManifestObject | ConvertTo-Manifest -Extension $Format
+
+    if ($output) {
+        Write-UserMessage -Message "Showing manifest for '$app'" -Success # TODO: Add better text with parsed appname, version, url/bucket
+        Write-UserMessage -Message $output -Output
     }
 }