+++ supporting/completion/scoop.lua
@@ -14,6 +14,7 @@ end
 local booleanParser = parser({'true', 'false'})
 local architectureParser = parser({'32bit', '64bit'})
 local utilityParser = parser({'native', 'aria2'})
+local manifestFormatParser = parser({'json', 'yml', 'yaml'})
 local configOptions = parser({
     '7ZIPEXTRACT_USE_EXTERNAL' .. booleanParser,
     'aria2-enabled' .. booleanParser,
@@ -152,6 +153,7 @@ local scoopParser = parser({
         '-h', '--help'
     }),
     'cat' .. parser({getLocallyAvailableApplicationsByScoop},
+        '-f' .. manifestFormatParser, '--format' .. manifestFormatParser,
         '-h', '--help'
     ),
     'cache' .. parser({'show', 'rm'} .. parser({getScoopCachedFile}),