+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.6.5](https://github.com/Ash258/Scoop-Core/milestone/5)
+
+- **scoop-cat**: Add `-f`, `--format` options
+- Adopt new resolve function for parameter passing
+    - **scoop-cat**
+    - **scoop-download**
+
 ## [0.6](https://github.com/Ash258/Scoop-Core/milestone/4)
 
 ### 0.6-pre4